Refactors dashboard state to a component and supporting service
In support of the longer term fix brought on by https://github.com/ManageIQ/manageiq-ui-service/pull/1372
