+++ client/app/components/_components.sass
@@ -3,3 +3,4 @@
 // @import all of your component styles here
 
 @import 'notifications/notifications'
+@import './dashboard/dashboard'
