+++ client/app/components/dashboard/dashboard.html
@@ -1,17 +1,17 @@
 <div class="container-fluid container-cards-pf ss-dashboard">
   <div class="row row-cards-pf">
-    <div class="col-xs-6 col-sm-6 col-md-6" ng-show="vm.servicesFeature">
+    <div class="col-xs-6 col-sm-6 col-md-6" ng-if="vm.servicesFeature">
       <div class="card-pf card-pf-accented ss-dashboard__card-primary" ui-sref="services">
         <div class="card-pf-body">
           <span class="fa fa-file-o"></span>
           <div class="ss-dashboard__card-primary__count">
-            <h2>{{ ::vm.servicesCount.total }}</h2>
+            <h2>{{ vm.servicesCount.total }}</h2>
             <h3 translate>Total Services</h3>
           </div>
         </div>
       </div>
     </div>
-    <div class="col-xs-6 col-sm-6 col-md-6" ng-show="vm.requestsFeature">
+    <div class="col-xs-6 col-sm-6 col-md-6" ng-if="vm.requestsFeature">
       <div class="card-pf card-pf-accented ss-dashboard__card-primary total-requests">
         <div class="card-pf-body">
           <span class="fa fa-file-text-o"></span>
@@ -24,7 +24,7 @@
     </div>
   </div>
   <div class="row row-cards-pf">
-    <div class="col-xs-6 col-sm-6 col-md-6" ng-show="vm.servicesFeature">
+    <div class="col-xs-6 col-sm-6 col-md-6" ng-if="vm.servicesFeature">
       <div class="card-pf">
         <div class="row row-cards-pf">
           <div class="card-pf card-pf-aggregate-status card-divider">
@@ -37,7 +37,7 @@
                   <span class="card-pf-aggregate-status-notification">
                     <span class="pficon pficon-warning-triangle-o" uib-tooltip="{{'The number of services retiring within the next 30 days'|translate}}" tooltip-placement="bottom"></span>
                   </span>
-                  <span class="card-pf-aggregate-status-notification">{{ ::vm.servicesCount.soon }}</span>
+                  <span class="card-pf-aggregate-status-notification">{{ vm.servicesCount.soon }}</span>
                 </p>
               </div>
             </a>
@@ -52,7 +52,7 @@
                   <span class="card-pf-aggregate-status-notification">
                     <span class="pficon fa fa-check" uib-tooltip="{{'The number of active services'|translate}}" tooltip-placement="bottom"></span>
                   </span>
-                  <span class="card-pf-aggregate-status-notification">{{ ::vm.servicesCount.current }}</span>
+                  <span class="card-pf-aggregate-status-notification">{{ vm.servicesCount.current }}</span>
                 </p>
               </div>
             </a>
@@ -67,7 +67,7 @@
                   <span class="card-pf-aggregate-status-notification">
                     <span class="pficon pficon-close" uib-tooltip="{{'The number of services which have hit their retirement period or been retired'|translate}}" tooltip-placement="bottom"></span>
                   </span>
-                  <span class="card-pf-aggregate-status-notification">{{ ::vm.servicesCount.retired }}
+                  <span class="card-pf-aggregate-status-notification">{{ vm.servicesCount.retired }}
                   </span>
                 </p>
               </div>
@@ -94,47 +94,47 @@
         </div>
       </div>
     </div>
-    <div class="col-xs-6 col-sm-6 col-md-6" ng-show="vm.requestsFeature">
+    <div class="col-xs-6 col-sm-6 col-md-6" ng-if="vm.requestsFeature">
       <div class="card-pf">
         <div class="row row-cards-pf">
           <div class="card-pf card-pf-aggregate-status card-divider">
-              <div class="card-pf-body">
-                <h2 class="card-pf-title" translate>
-                  Pending Requests
-                </h2>
-                <p class="card-pf-aggregate-status-notifications">
+            <div class="card-pf-body">
+              <h2 class="card-pf-title" translate>
+                Pending Requests
+              </h2>
+              <p class="card-pf-aggregate-status-notifications">
                   <span class="card-pf-aggregate-status-notification">
                     <span class="pficon pficon-warning-triangle-o"></span>
                   </span>
-                  <span class="card-pf-aggregate-status-notification">{{ ::vm.requestsCount.pending }}</span>
-                </p>
-              </div>
+                <span class="card-pf-aggregate-status-notification">{{ vm.requestsCount.pending }}</span>
+              </p>
+            </div>
           </div>
           <div class="card-pf card-pf-aggregate-status card-divider">
-              <div class="card-pf-body">
-                <h2 class="card-pf-title" translate>
-                  Approved Requests
-                </h2>
-                <p class="card-pf-aggregate-status-notifications">
+            <div class="card-pf-body">
+              <h2 class="card-pf-title" translate>
+                Approved Requests
+              </h2>
+              <p class="card-pf-aggregate-status-notifications">
                   <span class="card-pf-aggregate-status-notification">
                     <span class="pficon fa fa-check"></span>
                   </span>
-                  <span class="card-pf-aggregate-status-notification">{{ ::vm.requestsCount.approved }}</span>
-                </p>
-              </div>
+                <span class="card-pf-aggregate-status-notification">{{ vm.requestsCount.approved }}</span>
+              </p>
+            </div>
           </div>
           <div class="card-pf card-pf-aggregate-status card-divider">
-              <div class="card-pf-body">
-                <h2 class="card-pf-title" translate>
-                  Denied Requests
-                </h2>
-                <p class="card-pf-aggregate-status-notifications">
+            <div class="card-pf-body">
+              <h2 class="card-pf-title" translate>
+                Denied Requests
+              </h2>
+              <p class="card-pf-aggregate-status-notifications">
                   <span class="card-pf-aggregate-status-notification">
                     <span class="pficon pficon pficon-close"></span>
                   </span>
-                  <span class="card-pf-aggregate-status-notification">{{ ::vm.requestsCount.denied }}</span>
-                </p>
-              </div>
+                <span class="card-pf-aggregate-status-notification">{{ vm.requestsCount.denied }}</span>
+              </p>
+            </div>
           </div>
         </div>
       </div>
