+++ client/app/states/_states.sass
@@ -9,5 +9,4 @@
 @import '404/404'
 @import 'login/login'
 @import 'catalogs/details/details'
-@import 'dashboard/dashboard'
 @import 'orders/details/details'
